ML-Image-Analyzer/ML-Image-Analyzer/requirements.txt
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.0.0  # This is likely needed for color analysis
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.22.0
scipy>=1.11.0 